tests: name the remaining test files after their component, not the incident that produced them
Second and final pass of the reorganisation begun in 6179ce90. Files named for a
review, phase, or aspect ("_extended", "_wireup", "_wired", "_audit_status",
"_safe_get", "_branch") say why they were written rather than what they cover,
which is useless to anyone reading them later. Each folds into the component
file that already existed.

Merged away (source -> target):
  core/test_config_base_extended            -> core/test_config_base
  core/test_config_manager_extended         -> core/test_config_manager
  core/test_config_version_extended         -> core/test_config_version
  core/test_state_manager_extended          -> core/test_state_manager
  data/test_credential_store_extended       -> data/test_credential_store
  data/test_database_pooling_extended       -> data/test_database_pooling
  data/test_conversion_utils_safe_get       -> data/test_conversion_utils
  execution/test_execution_engine_extended  -> execution/test_execution_engine
  optimization/test_dynamic_optimizer_extended -> optimization/test_dynamic_optimizer
  portfolio/test_portfolio_manager_extended -> portfolio/test_portfolio_manager
  risk/test_risk_manager_extended           -> risk/test_risk_manager
  instruments/test_equity_instrument_holiday_check_wired -> instruments/test_equity
  instruments/test_equity_audit_status      -> instruments/test_equity
  instruments/test_equity_leverage_guardrail -> instruments/test_equity
  instruments/test_exchange_json_wireup     -> instruments/test_instrument_registry
  strategy/test_pnl_accounting_branch       -> strategy/test_base_strategy

Renamed (no component file existed to merge into):
  transaction_cost/test_adv_classifier_wireup -> test_transaction_cost_manager
  core/test_holiday_lookback_extended_closure -> core/test_holiday_lookback

Folded sections are wrapped in named namespaces because several files defined
colliding helpers; gtest identity is the suite+name string, so no case was
renamed or disabled. Verified zero suite+name collisions before merging.
1386 TEST/TEST_F declarations before and after; suite 1431/1431 unchanged.

Two merge hazards found and fixed rather than worked around:
  - test_risk_manager reaches RiskManager private members. The source guarded
    its include with `#define private public` AFTER pre-loading every header
    risk_manager.hpp transitively pulls in, so the macro only affects that one
    header. Hoisting includes broke both halves of that: the guard no longer
    wrapped the include, and once restored too early it leaked into libc++.
    Include order is load-bearing here and is now commented as such.
  - test_base_strategy declares a file-scope MockPostgresDatabase while the
    folded tests were written against trade_ngin::testing's. Inside the
    namespace both were visible; the original resolution is now pinned by alias.

test_extended_kalman_filter is untouched -- "Extended Kalman Filter" is the
algorithm's name, not a phase label.
