+++ tests/CMakeLists.txt
@@ -3,33 +3,25 @@ set(TEST_SOURCES
     core/test_decimal.cpp
     core/test_result.cpp
     core/test_state_manager.cpp
-    core/test_state_manager_extended.cpp
     core/test_logger.cpp
     core/test_config_manager.cpp
     core/test_config_version.cpp
-    core/test_config_version_extended.cpp
     core/test_config_base.cpp
-    core/test_holiday_lookback_extended_closure.cpp
-    core/test_config_base_extended.cpp
+    core/test_holiday_lookback.cpp
     core/test_run_id_generator.cpp
     core/test_config_loader.cpp
-    core/test_config_manager_extended.cpp
     data/test_db_utils.cpp
-    data/test_conversion_utils_safe_get.cpp
     data/test_utc_date_contract.cpp
     data/test_postgres_param_safety.cpp
     core/test_holiday_path_resolution.cpp
     core/test_holiday_calendar_coverage.cpp
-    instruments/test_equity_audit_status.cpp
     instruments/test_equity_market_hours_exchange.cpp
     data/test_postgres_database.cpp
     data/test_database_pooling.cpp
     data/test_db_transaction_atomicity.cpp
-    data/test_database_pooling_extended.cpp
     data/test_market_data_bus.cpp
     data/test_credential_store.cpp
     data/test_bar_close_uses_adjusted.cpp
-    data/test_credential_store_extended.cpp
     data/test_conversion_utils.cpp
     data/test_market_data_utils.cpp
     order/test_order_manager.cpp
@@ -39,11 +31,9 @@ set(TEST_SOURCES
     instruments/test_option.cpp
     instruments/test_instrument_registry.cpp
     risk/test_risk_manager.cpp
-    risk/test_risk_manager_extended.cpp
     optimization/test_dynamic_optimizer.cpp
-    optimization/test_dynamic_optimizer_extended.cpp
     backtest/test_transaction_cost_analysis.cpp
-    transaction_cost/test_adv_classifier_wireup.cpp
+    transaction_cost/test_transaction_cost_manager.cpp
     transaction_cost/test_unknown_equity_costs_dispatch.cpp
     transaction_cost/test_equity_borrow_fees.cpp
     transaction_cost/test_spread_model.cpp
@@ -54,7 +44,6 @@ set(TEST_SOURCES
     live/corp_actions/test_income.cpp
     live/test_live_metrics_includes_dividend_income.cpp
     live/test_daily_pnl_identity.cpp
-    strategy/test_pnl_accounting_branch.cpp
     backtest/test_slippage_model.cpp
     backtest/test_backtest_execution_manager.cpp
     backtest/test_backtest_price_manager.cpp
@@ -65,17 +54,12 @@ set(TEST_SOURCES
     backtest/test_backtest_csv_exporter.cpp
     backtest/test_backtest_coordinator.cpp
     execution/test_execution_engine.cpp
-    execution/test_execution_engine_extended.cpp
     portfolio/test_portfolio_manager.cpp
-    instruments/test_equity_instrument_holiday_check_wired.cpp
-    instruments/test_exchange_json_wireup.cpp
     core/test_equity_portfolio_namespace.cpp
     live/corp_actions/test_window.cpp
     instruments/test_equity_margin_cash_regt.cpp
-    instruments/test_equity_leverage_guardrail.cpp
     instruments/test_futures_margin_overload.cpp
     portfolio/test_portfolio_aggregation.cpp
-    portfolio/test_portfolio_manager_extended.cpp
     portfolio/test_portfolio_manager_internals.cpp
     storage/test_backtest_results_manager.cpp
     storage/test_live_results_manager.cpp
