+++ tests/core/test_holiday_lookback.cpp
