--- tests/execution/test_execution_engine_extended.cpp
@@ -1,214 +1,0 @@
-// Extended branch coverage for execution_engine.cpp. Targets accessor error
-// paths, custom-algo dispatch, repeated-initialization, get_active_jobs.
-
-#include <gtest/gtest.h>
-#include <atomic>
-#include <chrono>
-#include <memory>
-#include <thread>
-#include "../core/test_base.hpp"
-#include "../order/test_utils.hpp"
-#include "trade_ngin/execution/execution_engine.hpp"
-#include "trade_ngin/order/order_manager.hpp"
-
-using namespace trade_ngin;
-using namespace trade_ngin::testing;
-
-class ExecutionEngineExtendedTest : public TestBase {
-protected:
-    void SetUp() override {
-        TestBase::SetUp();
-        OrderManagerConfig oc = create_test_config();
-        oc.simulate_fills = true;
-        oc.max_notional_value = 2'000'000.0;
-        order_manager_ = std::make_shared<OrderManager>(oc);
-        ASSERT_TRUE(order_manager_->initialize().is_ok());
-        engine_ = std::make_unique<ExecutionEngine>(order_manager_);
-        ASSERT_TRUE(engine_->initialize().is_ok());
-    }
-
-    void TearDown() override {
-        if (engine_) {
-            auto jobs = engine_->get_active_jobs();
-            if (jobs.is_ok()) {
-                for (const auto& job : jobs.value()) {
-                    engine_->cancel_execution(job.job_id);
-                }
-            }
-            std::this_thread::sleep_for(std::chrono::milliseconds(50));
-            engine_.reset();
-        }
-        order_manager_.reset();
-        TestBase::TearDown();
-    }
-
-    Order make_order(const std::string& symbol, Side side, double qty, double price) {
-        Order o;
-        o.symbol = symbol;
-        o.side = side;
-        o.quantity = qty;
-        o.price = price;
-        o.type = OrderType::MARKET;
-        o.time_in_force = TimeInForce::DAY;
-        return o;
-    }
-
-    ExecutionConfig basic_config() {
-        ExecutionConfig c;
-        c.max_participation_rate = 0.1;
-        c.urgency_level = 0.5;
-        c.time_horizon = std::chrono::minutes(60);
-        return c;
-    }
-
-    std::shared_ptr<OrderManager> order_manager_;
-    std::unique_ptr<ExecutionEngine> engine_;
-};
-
-// ===== initialize is not idempotent =====
-
-TEST_F(ExecutionEngineExtendedTest, InitializeTwiceReturnsError) {
-    // FIXME: production behavior — ExecutionEngine::initialize() is NOT
-    // idempotent. The fixture's SetUp already initialized once; a second call
-    // returns an error (likely due to duplicate event-bus subscription
-    // registration). Captured here so the test fires if/when this becomes
-    // idempotent.
-    EXPECT_TRUE(engine_->initialize().is_error());
-}
-
-// ===== cancel_execution error path =====
-
-TEST_F(ExecutionEngineExtendedTest, CancelNonExistentJobReturnsError) {
-    auto r = engine_->cancel_execution("nonexistent-job-id");
-    EXPECT_TRUE(r.is_error());
-}
-
-// ===== get_metrics error path =====
-
-TEST_F(ExecutionEngineExtendedTest, GetMetricsForUnknownJobReturnsError) {
-    auto r = engine_->get_metrics("nonexistent-job-id");
-    EXPECT_TRUE(r.is_error());
-}
-
-// ===== get_active_jobs =====
-
-TEST_F(ExecutionEngineExtendedTest, GetActiveJobsEmptyBeforeAnySubmission) {
-    auto r = engine_->get_active_jobs();
-    ASSERT_TRUE(r.is_ok());
-    EXPECT_TRUE(r.value().empty());
-}
-
-TEST_F(ExecutionEngineExtendedTest, GetActiveJobsContainsSubmittedJob) {
-    auto sub = engine_->submit_execution(make_order("AAPL", Side::BUY, 10, 150.0),
-                                          ExecutionAlgo::TWAP, basic_config());
-    ASSERT_TRUE(sub.is_ok());
-    auto jobs = engine_->get_active_jobs();
-    ASSERT_TRUE(jobs.is_ok());
-    bool found = false;
-    for (const auto& job : jobs.value()) {
-        if (job.job_id == sub.value()) found = true;
-    }
-    EXPECT_TRUE(found);
-}
-
-TEST_F(ExecutionEngineExtendedTest, CancelExecutionRemovesJobFromActive) {
-    auto sub = engine_->submit_execution(make_order("AAPL", Side::BUY, 10, 150.0),
-                                          ExecutionAlgo::TWAP, basic_config());
-    ASSERT_TRUE(sub.is_ok());
-    EXPECT_TRUE(engine_->cancel_execution(sub.value()).is_ok());
-}
-
-TEST_F(ExecutionEngineExtendedTest, GetMetricsAvailableForSubmittedJob) {
-    auto sub = engine_->submit_execution(make_order("AAPL", Side::BUY, 10, 150.0),
-                                          ExecutionAlgo::TWAP, basic_config());
-    ASSERT_TRUE(sub.is_ok());
-    auto m = engine_->get_metrics(sub.value());
-    EXPECT_TRUE(m.is_ok());
-}
-
-// ===== custom algo registration =====
-
-TEST_F(ExecutionEngineExtendedTest, RegisterCustomAlgoSucceedsWithUniqueName) {
-    auto r = engine_->register_custom_algo(
-        "MY_CUSTOM_ALGO",
-        [](const ExecutionJob&) { return Result<void>(); });
-    EXPECT_TRUE(r.is_ok());
-}
-
-TEST_F(ExecutionEngineExtendedTest, RegisterCustomAlgoOverwritesSameName) {
-    std::atomic<int> v1_calls{0};
-    auto r1 = engine_->register_custom_algo(
-        "DUP_ALGO",
-        [&v1_calls](const ExecutionJob&) {
-            ++v1_calls;
-            return Result<void>();
-        });
-    EXPECT_TRUE(r1.is_ok());
-
-    // Re-register with the same name → either overwrites or rejects; we just
-    // exercise both code paths.
-    auto r2 = engine_->register_custom_algo(
-        "DUP_ALGO",
-        [](const ExecutionJob&) { return Result<void>(); });
-    // The contract may be either overwrite-ok or reject; both are acceptable
-    // structurally. Important is that it doesn't crash.
-    (void)r2;
-}
-
-// ===== submit_execution error paths =====
-
-TEST_F(ExecutionEngineExtendedTest, SubmitExecutionWithZeroQuantityHandled) {
-    Order o = make_order("AAPL", Side::BUY, 0.0, 150.0);
-    auto sub = engine_->submit_execution(o, ExecutionAlgo::TWAP, basic_config());
-    // Either rejected (preferred) or accepted but with no fills; both paths exercised.
-    (void)sub;
-}
-
-TEST_F(ExecutionEngineExtendedTest, SubmitExecutionWithZeroTimeHorizonHandled) {
-    auto cfg = basic_config();
-    cfg.time_horizon = std::chrono::minutes(0);
-    auto sub = engine_->submit_execution(make_order("AAPL", Side::BUY, 10, 150.0),
-                                          ExecutionAlgo::TWAP, cfg);
-    // Production may or may not validate this; structurally either result acceptable
-    if (sub.is_ok()) {
-        EXPECT_FALSE(sub.value().empty());
-    }
-}
-
-// ===== Multi-job tracking =====
-
-TEST_F(ExecutionEngineExtendedTest, RapidSubmissionsCollideOnJobIdAndOverwrite) {
-    // FIXME: production bug — submit_execution generates a job ID that includes
-    // a timestamp at coarse resolution, so two submissions in the same tick
-    // produce identical IDs. The second submission silently overwrites the
-    // first in active_jobs_, so only one job is tracked. Captured here so the
-    // test fires if/when ID generation becomes monotonic/unique.
-    auto a = engine_->submit_execution(make_order("AAPL", Side::BUY, 10, 150.0),
-                                         ExecutionAlgo::TWAP, basic_config());
-    auto b = engine_->submit_execution(make_order("MSFT", Side::BUY, 5, 250.0),
-                                         ExecutionAlgo::VWAP, basic_config());
-    ASSERT_TRUE(a.is_ok());
-    ASSERT_TRUE(b.is_ok());
-    // Currently a == b (collision). Either way, get_active_jobs reports at most
-    // however many distinct IDs were generated.
-    auto jobs = engine_->get_active_jobs();
-    ASSERT_TRUE(jobs.is_ok());
-    EXPECT_LE(jobs.value().size(), 2u);
-    EXPECT_GE(jobs.value().size(), 1u);
-}
-
-TEST_F(ExecutionEngineExtendedTest, SpacedSubmissionsAreTrackedDistinctly) {
-    // Sleep between submissions to dodge the same-tick ID collision.
-    auto a = engine_->submit_execution(make_order("AAPL", Side::BUY, 10, 150.0),
-                                         ExecutionAlgo::TWAP, basic_config());
-    std::this_thread::sleep_for(std::chrono::milliseconds(2));
-    auto b = engine_->submit_execution(make_order("MSFT", Side::BUY, 5, 250.0),
-                                         ExecutionAlgo::VWAP, basic_config());
-    ASSERT_TRUE(a.is_ok());
-    ASSERT_TRUE(b.is_ok());
-    if (a.value() != b.value()) {
-        auto jobs = engine_->get_active_jobs();
-        ASSERT_TRUE(jobs.is_ok());
-        EXPECT_GE(jobs.value().size(), 2u);
-    }
-}
