+++ tests/instruments/test_equity.cpp
@@ -3,6 +3,19 @@
 #include "../core/test_base.hpp"
 #include "trade_ngin/instruments/equity.hpp"
 
+#include <ctime>
+#include <filesystem>
+#include <fstream>
+#include <memory>
+#include "trade_ngin/core/holiday_checker.hpp"
+#include <gtest/gtest.h>
+
+#include <memory>
+
+#include "../data/test_db_utils.hpp"
+#include "trade_ngin/instruments/equity.hpp"
+#include "trade_ngin/instruments/instrument_registry.hpp"
+
 using namespace trade_ngin;
 using namespace trade_ngin::testing;
 
@@ -171,3 +184,289 @@ TEST_F(EquityInstrumentTest, ETFFlagPropagates) {
     EquityInstrument eq("SPY", spec);
     EXPECT_TRUE(eq.is_etf());
 }
+
+// ===== folded in from tests/instruments/test_equity_instrument_holiday_check_wired.cpp =====
+namespace equity_instrument_holiday_check_wired_detail {
+
+using namespace trade_ngin;
+
+// Regression test for audit finding §1.10: EquityInstrument::is_market_open()
+// was wired to consult HolidayChecker only via a static slot that was never
+// populated in production -- so it silently allowed Christmas-day trades.
+// Phase 3d calls EquityInstrument::set_holiday_checker() at app startup; this
+// test verifies the wired path actually rejects holiday timestamps.
+
+namespace {
+
+std::chrono::system_clock::time_point make_local(int year, int month, int day,
+                                                 int hour, int minute) {
+    std::tm tm{};
+    tm.tm_year = year - 1900;
+    tm.tm_mon = month - 1;
+    tm.tm_mday = day;
+    tm.tm_hour = hour;
+    tm.tm_min = minute;
+    tm.tm_isdst = -1;
+    auto t = std::mktime(&tm);
+    return std::chrono::system_clock::from_time_t(t);
+}
+
+class EquityInstrumentHolidayTest : public ::testing::Test {
+protected:
+    void SetUp() override {
+        tmp_dir_ = std::filesystem::temp_directory_path() /
+                   ("equity_holiday_" + std::to_string(std::rand()));
+        std::filesystem::create_directories(tmp_dir_);
+        holidays_path_ = tmp_dir_ / "holidays.json";
+
+        std::ofstream out(holidays_path_);
+        out << R"({
+  "2026": [
+    {"date": "2026-12-25", "name": "Christmas Day", "type": "federal"}
+  ]
+})";
+        out.close();
+
+        spec_.exchange = "NASDAQ";
+        spec_.currency = "USD";
+        spec_.tick_size = 0.01;
+        spec_.trading_hours = "09:30-16:00";
+
+        // Make a fresh checker and wire it into the static slot.
+        auto checker = std::make_shared<HolidayChecker>(holidays_path_.string());
+        EquityInstrument::set_holiday_checker(checker);
+    }
+
+    void TearDown() override {
+        // Reset the static so subsequent tests aren't surprised by our mock.
+        EquityInstrument::set_holiday_checker(nullptr);
+        std::error_code ec;
+        std::filesystem::remove_all(tmp_dir_, ec);
+    }
+
+    std::filesystem::path tmp_dir_;
+    std::filesystem::path holidays_path_;
+    EquitySpec spec_;
+};
+
+}  // namespace
+
+// Friday 2026-12-25 is Christmas (configured holiday). Even at 10:00 local
+// (well inside the 09:30-16:00 trading window) and a weekday, is_market_open
+// must report false because the holiday check fires.
+TEST_F(EquityInstrumentHolidayTest, ChristmasReturnsClosedWhenCheckerWired) {
+    EquityInstrument aapl("AAPL", spec_);
+    auto christmas_10am = make_local(2026, 12, 25, 10, 0);
+    EXPECT_FALSE(aapl.is_market_open(christmas_10am));
+}
+
+// Sanity: an ordinary Wednesday at 10am is open.
+TEST_F(EquityInstrumentHolidayTest, RegularWeekdayWithinHoursReturnsOpen) {
+    EquityInstrument aapl("AAPL", spec_);
+    auto wed_10am = make_local(2026, 12, 9, 10, 0);  // Wednesday
+    EXPECT_TRUE(aapl.is_market_open(wed_10am));
+}
+
+// If no checker is wired (e.g. pre-fix or after TearDown), the method falls
+// back to weekday-and-hours only. This documents the contract: the checker is
+// the holiday-aware path; without it, holidays slip through.
+TEST_F(EquityInstrumentHolidayTest, WithoutCheckerHolidayWouldBeMissed) {
+    EquityInstrument::set_holiday_checker(nullptr);
+    EquityInstrument aapl("AAPL", spec_);
+    auto christmas_10am = make_local(2026, 12, 25, 10, 0);
+    // Christmas 2026 is a Friday -- weekday, within hours -- so without a
+    // checker, the method incorrectly returns true. This is the pre-fix bug,
+    // pinned here as a contract assertion (NOT a desired behavior).
+    EXPECT_TRUE(aapl.is_market_open(christmas_10am))
+        << "Without a holiday checker, is_market_open should fall through to "
+           "weekday+hours only. If this fails, the fallback contract changed.";
+}
+
+}  // namespace equity_instrument_holiday_check_wired_detail
+
+// ===== folded in from tests/instruments/test_equity_audit_status.cpp =====
+namespace equity_audit_status_detail {
+
+using namespace trade_ngin;
+
+// Phase 6 §1.10 -- pin the closed-status of three audit findings the audit
+// originally flagged as "dead methods":
+//
+//   1. set_holiday_checker IS called by production apps -> the static slot is
+//      populated and is_market_open consults the calendar. We can't easily
+//      assert "the live app calls it" from a unit test, but we CAN assert
+//      that wiring a checker affects is_market_open in the expected way.
+//   2. calculate_commission returns qty * spec_.commission_per_share with a
+//      non-zero default tier ($0.005/share via instrument_registry.cpp:227).
+//      We pin the formula here; the registry test pins the population.
+//   3. get_margin_requirement(price, quantity) uses the Phase 2 §1.3 Reg T
+//      math (CASH 100% notional, REG_T long 50%, REG_T short 150%).
+//
+// If a future refactor accidentally re-introduces the dead-code state
+// (commission hardcoded to 0, holiday checker never consulted, margin
+// stuck at the 0.0 sentinel), these tests will fire.
+
+namespace {
+
+EquitySpec make_spec(double commission_per_share = 0.005,
+                     EquityAccountMode mode = EquityAccountMode::CASH) {
+    EquitySpec s;
+    s.exchange = "NYSE";
+    s.currency = "USD";
+    s.commission_per_share = commission_per_share;
+    s.account_mode = mode;
+    return s;
+}
+
+}  // namespace
+
+// §1.10 finding 2: commission formula, not a hardcoded 0.
+TEST(EquityAuditStatus, CommissionUsesPerShareRate) {
+    EquityInstrument inst("AAPL", make_spec(/*commission_per_share=*/0.005));
+    EXPECT_DOUBLE_EQ(inst.calculate_commission(100.0), 0.50);
+    EXPECT_DOUBLE_EQ(inst.calculate_commission(-100.0), 0.50);  // abs
+}
+
+// §1.10 finding 2 cont.: when the spec carries no commission, the formula
+// returns 0 -- which is the FALLBACK semantic, not the BROKEN semantic.
+// The bug (audit's claim) was that the method ALWAYS returned 0; the test
+// above proves it doesn't.
+TEST(EquityAuditStatus, CommissionZeroWhenSpecZero) {
+    EquityInstrument inst("FOO", make_spec(/*commission_per_share=*/0.0));
+    EXPECT_DOUBLE_EQ(inst.calculate_commission(100.0), 0.0);
+}
+
+// §1.10 finding 3 / Phase 2 §1.3: REG_T long position posts 50% margin.
+TEST(EquityAuditStatus, MarginRegTLong) {
+    EquityInstrument inst("AAPL", make_spec(0.005, EquityAccountMode::REG_T));
+    // 100 shares * $150 = $15,000 notional -> 50% = $7,500.
+    EXPECT_DOUBLE_EQ(inst.get_margin_requirement(150.0, 100.0), 7500.0);
+}
+
+// §1.10 finding 3 / Phase 2 §1.3: REG_T short position posts 150% (100% cash
+// proceeds collateral + 50% maintenance).
+TEST(EquityAuditStatus, MarginRegTShort) {
+    EquityInstrument inst("AAPL", make_spec(0.005, EquityAccountMode::REG_T));
+    EXPECT_DOUBLE_EQ(inst.get_margin_requirement(150.0, -100.0), 22500.0);
+}
+
+// §1.10 finding 3 / Phase 2 §1.3: CASH account posts 100% notional.
+TEST(EquityAuditStatus, MarginCashFullNotional) {
+    EquityInstrument inst("AAPL", make_spec(0.005, EquityAccountMode::CASH));
+    EXPECT_DOUBLE_EQ(inst.get_margin_requirement(150.0, 100.0), 15000.0);
+}
+
+// §1.10 finding 1: wiring a holiday checker makes is_market_open consult
+// the calendar. Without the checker, the method falls through to
+// weekday/trading-hours logic only. Phase 6 §6b also pins the
+// thread-safety wrapper via atomic_load/store, but the assertion here is
+// purely "the checker is honored when set."
+//
+// We assert the negative case to avoid relying on a specific real holiday
+// JSON: register an empty checker, and a regular Wednesday afternoon
+// should still be "open" (weekday + within trading hours), demonstrating
+// the call-through is alive.
+TEST(EquityAuditStatus, HolidayCheckerIsConsulted) {
+    // Wire a checker (file may or may not exist; load failure is logged
+    // but doesn't throw).
+    auto checker = std::make_shared<HolidayChecker>(
+        HolidayChecker::resolve_holidays_path());
+    EquityInstrument::set_holiday_checker(checker);
+    auto roundtrip = EquityInstrument::get_holiday_checker();
+    EXPECT_EQ(roundtrip.get(), checker.get())
+        << "set/get holiday_checker must round-trip via the atomic slot";
+}
+
+}  // namespace equity_audit_status_detail
+
+// ===== folded in from tests/instruments/test_equity_leverage_guardrail.cpp =====
+namespace equity_leverage_guardrail_detail {
+
+using namespace trade_ngin;
+using namespace trade_ngin::testing;
+
+// Audit test T2.4 -- leverage guardrail (audit §3.3). The equity apps refuse
+// to start if max_gross_leverage > 1.0 and any CASH-mode equity is in the
+// symbol list. Cash accounts can't borrow, so a leverage cap above 1.0 is
+// structurally invalid.
+//
+// The guard logic lives inline in the apps' main(); this test verifies the
+// building blocks (instrument account_mode + registry lookup) the guard
+// depends on, plus reproduces the same check pattern the apps execute.
+
+namespace {
+
+// Mimic the exact check the equity apps perform after load_equity_instruments.
+// Returns true if the guardrail should refuse startup.
+bool would_refuse_startup(const std::vector<std::string>& symbols,
+                          const InstrumentRegistry& registry,
+                          double max_gross_leverage) {
+    if (max_gross_leverage <= 1.0) return false;
+    for (const auto& symbol : symbols) {
+        auto inst = registry.get_equity_instrument(symbol);
+        if (inst && inst->get_account_mode() == EquityAccountMode::CASH) {
+            return true;
+        }
+    }
+    return false;
+}
+
+class LeverageGuardrailTest : public TestBase {
+protected:
+    void SetUp() override {
+        TestBase::SetUp();
+        auto& registry = InstrumentRegistry::instance();
+        auto db = std::make_shared<MockPostgresDatabase>("mock://leverage_guardrail_test");
+        ASSERT_TRUE(db->connect().is_ok());
+        (void)registry.initialize(db);  // Idempotent.
+    }
+};
+
+}  // namespace
+
+TEST_F(LeverageGuardrailTest, CashEquityWithLeverageAboveOneRefuses) {
+    auto& registry = InstrumentRegistry::instance();
+    const std::string sym = "PHASE2_T24_CASH_AAPL";
+
+    EquitySpec spec;
+    spec.exchange = "NASDAQ";
+    spec.currency = "USD";
+    spec.tick_size = 0.01;
+    spec.account_mode = EquityAccountMode::CASH;
+    registry.register_instrument(sym, std::make_shared<EquityInstrument>(sym, spec));
+
+    EXPECT_TRUE(would_refuse_startup({sym}, registry, 2.0))
+        << "Cash equity with max_gross_leverage=2.0 must trip the guardrail.";
+}
+
+TEST_F(LeverageGuardrailTest, CashEquityWithLeverageOneAccepts) {
+    auto& registry = InstrumentRegistry::instance();
+    const std::string sym = "PHASE2_T24_CASH_AAPL_LEV1";
+
+    EquitySpec spec;
+    spec.exchange = "NASDAQ";
+    spec.currency = "USD";
+    spec.tick_size = 0.01;
+    spec.account_mode = EquityAccountMode::CASH;
+    registry.register_instrument(sym, std::make_shared<EquityInstrument>(sym, spec));
+
+    EXPECT_FALSE(would_refuse_startup({sym}, registry, 1.0))
+        << "Cash equity at max_gross_leverage=1.0 (the valid cap) must NOT trip the guardrail.";
+}
+
+TEST_F(LeverageGuardrailTest, RegTEquityWithLeverageAboveOneAccepts) {
+    auto& registry = InstrumentRegistry::instance();
+    const std::string sym = "PHASE2_T24_REGT_MSFT";
+
+    EquitySpec spec;
+    spec.exchange = "NASDAQ";
+    spec.currency = "USD";
+    spec.tick_size = 0.01;
+    spec.account_mode = EquityAccountMode::REG_T;
+    registry.register_instrument(sym, std::make_shared<EquityInstrument>(sym, spec));
+
+    EXPECT_FALSE(would_refuse_startup({sym}, registry, 2.0))
+        << "REG_T equities permit leverage > 1.0; guardrail must not fire.";
+}
+
+}  // namespace equity_leverage_guardrail_detail
