--- tests/instruments/test_equity_audit_status.cpp
@@ -1,96 +1,0 @@
-#include <gtest/gtest.h>
-
-#include <memory>
-
-#include "trade_ngin/core/holiday_checker.hpp"
-#include "trade_ngin/instruments/equity.hpp"
-
-using namespace trade_ngin;
-
-// Phase 6 §1.10 -- pin the closed-status of three audit findings the audit
-// originally flagged as "dead methods":
-//
-//   1. set_holiday_checker IS called by production apps -> the static slot is
-//      populated and is_market_open consults the calendar. We can't easily
-//      assert "the live app calls it" from a unit test, but we CAN assert
-//      that wiring a checker affects is_market_open in the expected way.
-//   2. calculate_commission returns qty * spec_.commission_per_share with a
-//      non-zero default tier ($0.005/share via instrument_registry.cpp:227).
-//      We pin the formula here; the registry test pins the population.
-//   3. get_margin_requirement(price, quantity) uses the Phase 2 §1.3 Reg T
-//      math (CASH 100% notional, REG_T long 50%, REG_T short 150%).
-//
-// If a future refactor accidentally re-introduces the dead-code state
-// (commission hardcoded to 0, holiday checker never consulted, margin
-// stuck at the 0.0 sentinel), these tests will fire.
-
-namespace {
-
-EquitySpec make_spec(double commission_per_share = 0.005,
-                     EquityAccountMode mode = EquityAccountMode::CASH) {
-    EquitySpec s;
-    s.exchange = "NYSE";
-    s.currency = "USD";
-    s.commission_per_share = commission_per_share;
-    s.account_mode = mode;
-    return s;
-}
-
-}  // namespace
-
-// §1.10 finding 2: commission formula, not a hardcoded 0.
-TEST(EquityAuditStatus, CommissionUsesPerShareRate) {
-    EquityInstrument inst("AAPL", make_spec(/*commission_per_share=*/0.005));
-    EXPECT_DOUBLE_EQ(inst.calculate_commission(100.0), 0.50);
-    EXPECT_DOUBLE_EQ(inst.calculate_commission(-100.0), 0.50);  // abs
-}
-
-// §1.10 finding 2 cont.: when the spec carries no commission, the formula
-// returns 0 -- which is the FALLBACK semantic, not the BROKEN semantic.
-// The bug (audit's claim) was that the method ALWAYS returned 0; the test
-// above proves it doesn't.
-TEST(EquityAuditStatus, CommissionZeroWhenSpecZero) {
-    EquityInstrument inst("FOO", make_spec(/*commission_per_share=*/0.0));
-    EXPECT_DOUBLE_EQ(inst.calculate_commission(100.0), 0.0);
-}
-
-// §1.10 finding 3 / Phase 2 §1.3: REG_T long position posts 50% margin.
-TEST(EquityAuditStatus, MarginRegTLong) {
-    EquityInstrument inst("AAPL", make_spec(0.005, EquityAccountMode::REG_T));
-    // 100 shares * $150 = $15,000 notional -> 50% = $7,500.
-    EXPECT_DOUBLE_EQ(inst.get_margin_requirement(150.0, 100.0), 7500.0);
-}
-
-// §1.10 finding 3 / Phase 2 §1.3: REG_T short position posts 150% (100% cash
-// proceeds collateral + 50% maintenance).
-TEST(EquityAuditStatus, MarginRegTShort) {
-    EquityInstrument inst("AAPL", make_spec(0.005, EquityAccountMode::REG_T));
-    EXPECT_DOUBLE_EQ(inst.get_margin_requirement(150.0, -100.0), 22500.0);
-}
-
-// §1.10 finding 3 / Phase 2 §1.3: CASH account posts 100% notional.
-TEST(EquityAuditStatus, MarginCashFullNotional) {
-    EquityInstrument inst("AAPL", make_spec(0.005, EquityAccountMode::CASH));
-    EXPECT_DOUBLE_EQ(inst.get_margin_requirement(150.0, 100.0), 15000.0);
-}
-
-// §1.10 finding 1: wiring a holiday checker makes is_market_open consult
-// the calendar. Without the checker, the method falls through to
-// weekday/trading-hours logic only. Phase 6 §6b also pins the
-// thread-safety wrapper via atomic_load/store, but the assertion here is
-// purely "the checker is honored when set."
-//
-// We assert the negative case to avoid relying on a specific real holiday
-// JSON: register an empty checker, and a regular Wednesday afternoon
-// should still be "open" (weekday + within trading hours), demonstrating
-// the call-through is alive.
-TEST(EquityAuditStatus, HolidayCheckerIsConsulted) {
-    // Wire a checker (file may or may not exist; load failure is logged
-    // but doesn't throw).
-    auto checker = std::make_shared<HolidayChecker>(
-        HolidayChecker::resolve_holidays_path());
-    EquityInstrument::set_holiday_checker(checker);
-    auto roundtrip = EquityInstrument::get_holiday_checker();
-    EXPECT_EQ(roundtrip.get(), checker.get())
-        << "set/get holiday_checker must round-trip via the atomic slot";
-}
